# back-end/music-selector/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "MusicSelector API",
    "description": "API de recomendação musical com IA - Sistema multiplataforma de recomendação musical com classificação por estado emocional e contexto situacional",
    "version": "1.0.0",
    "contact": {
      "name": "MusicSelector Team",
      "url": "https://musicSelector.com"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Development server"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
      "LoginRequest": {
        "type": "object",
        "required": ["email", "password"],
        "properties": {
          "email": {
            "type": "string",
            "format": "email",
            "example": "user@example.com"
          },
          "password": {
            "type": "string",
            "format": "password",
            "minLength": 8,
            "example": "securePassword123"
          }
        }
      },
      "LoginResponse": {
        "type": "object",
        "properties": {
          "access_token": {
            "type": "string",
            "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
          }
        }
      },
      "CreateUserRequest": {
        "type": "object",
        "required": ["name", "email", "emailConfirmation", "password", "passwordConfirmation", "dateOfBirth"],
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 50,
            "example": "João Silva"
          },
          "email": {
            "type": "string",
            "format": "email",
            "example": "joao@example.com"
          },
          "emailConfirmation": {
            "type": "string",
            "format": "email",
            "example": "joao@example.com"
          },
          "password": {
            "type": "string",
            "minLength": 8,
            "example": "MyPassword123!"
          },
          "passwordConfirmation": {
            "type": "string",
            "minLength": 8,
            "example": "MyPassword123!"
          },
          "dateOfBirth": {
            "type": "string",
            "format": "date",
            "example": "2005-01-15"
          }
        }
      },
      "User": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "message": {
            "type": "string"
          }
        }
      },
...
      "OnboardingRequest": {
        "type": "object",
        "required": ["favoriteGenres", "audioPreference"],
        "properties": {
          "favoriteGenres": {
            "type": "array",
            "minItems": 1,
            "maxItems": 5,
            "items": {
              "type": "string"
            },
            "example": ["Rock", "Pop", "Jazz"]
          },
          "audioPreference": {
            "type": "string",
            "enum": ["VOCAL", "INSTRUMENTAL"],
            "example": "VOCAL"
          }
        }
      }
    }
  },
  "paths": {
...
    "/users/register": {
      "post": {
        "tags": ["Users"],
        "summary": "Registrar novo usuário",
        "description": "RN01-RN06: Cria novo usuário com validações de email, senha e idade. Email deve ser único, senha 8+ chars, idade 13+ anos",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateUserRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Usuário criado com sucesso",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              }
            }
          },
          "400": {
            "description": "Dados inválidos"
          },
          "409": {
            "description": "Email já cadastrado"
          }
        }
      }
    },
    "/users/login": {
      "post": {
        "tags": ["Users"],
        "summary": "Login de usuário",
        "description": "RN07-RN09: Autentica usuário e retorna JWT token",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Login bem-sucedido",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LoginResponse"
                }
              }
            }
          },
          "401": {
            "description": "Credenciais inválidas"
          }
        }
      }
    },
    "/users/forgot-password": {
      "post": {
        "tags": ["Users"],
        "summary": "Solicitar reset de senha",
        "description": "RN08: Envia link de reset de senha para o email (segurança: não revela se email existe)",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "email": {
                    "type": "string",
                    "format": "email"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Email de reset enviado (ou simulado)"
          }
        }
      }
    },
    "/users/reset-password": {
      "post": {
        "tags": ["Users"],
        "summary": "Resetar senha",
        "description": "RNF-S03: Reseta a senha com novo hash BCrypt",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "token": {
                    "type": "string"
                  },
                  "password": {
                    "type": "string",
                    "minLength": 8
                  },
                  "passwordConfirmation": {
                    "type": "string",
                    "minLength": 8
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Senha resetada com sucesso"
          },
          "400": {
            "description": "Senhas não correspondem"
          }
        }
      }
    },
    "/users/{id}/onboarding": {
      "post": {
        "tags": ["Users"],
        "summary": "Completar onboarding",
        "description": "RN10-RN13: Wizard de 3 passos (gêneros, estilo escuta, pref vocal/instrumental). Obrigatório na primeira vez",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "ID do usuário",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OnboardingRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Onboarding completado"
          }
        }
      }
    },
    "/users/{id}": {
      "patch": {
        "tags": ["Users"],
        "summary": "Atualizar perfil do usuário",
        "description": "Permite editar dados do perfil do usuário",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "ID do usuário",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Perfil atualizado"
          }
        }
      },
      "get": {
        "tags": ["Users"],
        "summary": "Obter dados do usuário",
        "description": "Retorna informações de um usuário específico",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "ID do usuário",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Dados do usuário",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              }
            }
          },
          "404": {
            "description": "Usuário não encontrado"
          }
        }
      },
      "delete": {
        "tags": ["Users"],
        "summary": "Deletar conta (soft delete)",
        "description": "RN29-RN30: LGPD - Deleta conta e anonimiza histórico de feedback",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "ID do usuário",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Conta deletada"
          }
        }
      }
    },
    "/users/{id}/hard": {
      "delete": {
        "tags": ["Users"],
        "summary": "Deletar conta permanentemente (hard delete)",
        "description": "Remove completamente o usuário e todos os dados",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "ID do usuário",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Usuário permanentemente deletado"
          }
        }
      }
    },
...
  },
...
}